+++ src/onegov.election_day/onegov/election_day/tests/test_layouts.py
@@ -93,7 +93,8 @@ def test_elections_layout():
     layout = ElectionsLayout(None, DummyRequest())
 
     assert layout.all_tabs == (
-        'lists', 'candidates', 'districts', 'connections', 'statistics', 'data'
+        'lists', 'candidates', 'districts', 'connections', 'statistics',
+        'panachage', 'data'
     )
 
     assert layout.title() == ''
@@ -104,6 +105,7 @@ def test_elections_layout():
     assert layout.title('connections') == 'List connections'
     assert layout.title('statistics') == 'Election statistics'
     assert layout.title('data') == 'Open Data'
+    assert layout.title('panachage') == 'Panachage'
 
     layout = ElectionsLayout(Election(type='majorz'), DummyRequest())
     assert layout.majorz